src/generativeUI/training-console/packages/api-server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-dotenv==1.0.1
slowapi>=0.1.9
# Test dependencies
anyio[trio]>=4.0.0
pytest-anyio>=0.0.0
